backend/pyproject.toml
[tool.poetry]
name = "ai-travel-planner-backend"
version = "0.1.0"
description = "FastAPI backend for AI Travel Planner"
authors = ["AI Travel Planner Team <dev@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
httpx = "^0.27.2"
openai = "^1.52.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
redis = "^5.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pytest = "^8.3.3"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
